fix: daemon Darwin build support
